requirements.txt
aniso8601==3.0.2
Django==2.2
django-extensions==2.1.6
django-versatileimagefield==1.10
docopt==0.6.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
jedi==0.13.3
parso==0.4.0
Pillow==5.4.1
promise==2.2.1
prompt-toolkit==2.0.9
ptpython==2.0.4
Pygments==2.3.1
pytz==2019.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
